requirements.txt
python-can
cantools
pylint
tabulate